edit all project

+++ Models/CaseProperties/Category.cs
@@ -3,11 +3,11 @@ using System.ComponentModel.DataAnnotations.Schema;
 
 namespace GraduationProject.Models.CaseProperties
 {
-	public class Category
-	{
-		public byte Id { get; set; }
+    public class Category
+    {
+        public int Id { get; set; }
 
-		[Required, MaxLength(50), Column(TypeName = "varchar")]
-		public string Name { get; set; }
-	}
+        [Required, MaxLength(50), Column(TypeName = "varchar")]
+        public string Name { get; set; }
+    }
 }
